+++ apis/userApi.ts
@@ -1,4 +1,5 @@
-import { fetchFormAPI } from "../utils/fetchFormAPI"
+import { fetchFormAPI } from "./api.utils"
+
 
 const USER_API_URL = "/posts"
