+++ app/(afterLogin)/createPost/_components/form/formSelector.tsx
@@ -1,9 +1,9 @@
 import { useCreatePostStore } from "@/libs/store"
-import { formatISODateString } from "@/utils/formatDate"
 import { FormSelectorProps } from "./type"
 import Image from "next/image"
 import { Theme } from "@/types/theme"
 import TextDisplay from "./formTextDisplay"
+import { formatISODateString } from "@/app/(afterLogin)/detailPost/[postId]/date.utils"
 
 const FormSelector = ({ onClick, label, state, postDetail, isThemeOpen, isTheme }: FormSelectorProps) => {
     const { selectedContinent, selectedCountry, startDate, endDate, selectedTheme, selectedAddress } =
