+++ app/(afterLogin)/user/[userId]/_components/myMap/worldMap.tsx
@@ -4,8 +4,8 @@ import React, { useEffect, useState, MouseEvent } from "react"
 import { WorldMapProps, Pin } from "./type"
 import { useMapStore } from "@/libs/pinStore"
 import { getPin, postPin } from "@/apis/mapApi"
-import { getPinLocalStorage, setPinLocalStorage } from "@/utils/localStorage"
 import Image from "next/image"
+import { getPinLocalStorage, setPinLocalStorage } from "@/utils/storage.utils"
 
 const WorldMap: React.FC<WorldMapProps> = ({ userInfo }) => {
     const { email, nickname } = userInfo
