+++ app/(beforeLogin)/search/page.tsx
@@ -4,7 +4,7 @@ import { getPost } from "@/apis/postApi"
 import SortDropdown from "@/components/commons/sortDropdown"
 import { useCreatePostStore } from "@/libs/store"
 import { Post } from "@/types/post"
-import { filterPosts } from "@/utils/filterPosts"
+import { filterPosts } from "@/utils/search.utils"
 import dynamic from "next/dynamic"
 import { useSearchParams } from "next/navigation"
 import { useEffect, useState } from "react"
