+++ constants/theme.ts
@@ -6,6 +6,6 @@ export const ThemeBannerContext: { [key: string]: string[] }[] = [
     { ACTIVITY: ["지루할 틈이 없는!", "액티비티 여행의", "기록을 확인할 수 있어요!"] },
     { SIGHTSEEING: ["여긴 꼭 가아해!", "관광명소 여행의", "기록을 확인할 수 있어요!"] },
     { PACKAGE: ["계획대로 척척!", "패키지 여행의", "기록을 확인할 수 있어요!"] },
-    { SUPER_LUXURY: ["고급, 고급, 고급!", "초호화 여행의", "기록을 확인할 수 있어요!"] },
-    { COST_EFFECTIVENESS: ["여행도 알뜰하게!", "가성비 여행의", "기록을 확인할 수 있어요!"] },
+    { LUXURY: ["고급, 고급, 고급!", "초호화 여행의", "기록을 확인할 수 있어요!"] },
+    { COST_SAVING: ["여행도 알뜰하게!", "가성비 여행의", "기록을 확인할 수 있어요!"] },
 ]
