+++ utils/resizeImage.utils.ts
